Rename package from com.greeting.demo to com.sinemergingtechnologies.database
