+++ src/main/java/com/sinemergingtechnologies/database/Client.java
@@ -1,4 +1,4 @@
-package com.greeting.demo;
+package com.sinemergingtechnologies.database;
 
 import lombok.*;
